requirements.txt
pandas 
numpy
nltk
beautifulsoup4
PyYAML
